Mega refactORE
