+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>org.openredstone</groupId>
     <artifactId>NetworkUnifier</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>2.0</version>
 
     <repositories>
         <repository>
