+++ src/main/resources/plugin.yml
@@ -1,4 +1,4 @@
 name: NetworkUnifier
 main: org.openredstone.NetworkUnifier
-version: 1.0
+version: 2.0
 author: Nickster258
